refactored contract details component
